added music and instructions

+++ GDD_2019_GROUP14/Assets/Scripts/MovementScript.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 public class MovementScript : MonoBehaviour
 {
     // speed
-    public int speed = 8;
+    public float speed = 8;
     public Vector3 orientation;
     // Start is called before the first frame update
     void Start()
